perf(daw-mixer): code-split desktop islands via next/dynamic — exclude chunk from mobile
